Dang, I havent committed in a while

+++ src/com/njdaeger/java/wrapper/User.java
@@ -488,9 +488,9 @@ public final class User implements IUser {
 			exists = true;
 		}
 		if (memory) {
-			return (double) Transform.getValue(player, PlayerPaths.FLYSPEED);
+			return (int) Transform.getValue(player, PlayerPaths.FLYSPEED);
 		}
-		return (double) userFile.getValue(PlayerPaths.FLYSPEED.getPath());
+		return (int) userFile.getValue(PlayerPaths.FLYSPEED.getPath());
 	}
 
 	@Override
